move buisness logic to service file
